refactor PlaceStorage & update mobx version
